--- a/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/FQDN.php
+++ b/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/FQDN.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace FQDNNamespacePrefix;
+
+class FQDN {}