--- a/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/FQDN/SomeClass.php
+++ b/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/FQDN/SomeClass.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace FQDNNamespacePrefix\FQDN;
+
+class SomeClass {}