--- a/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/Uses/Foo.php
+++ b/tests/DependencyEmitter/Fixtures/FQDNNamespacePrefix/Uses/Foo.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace FQDNNamespacePrefix\Uses;
+
+use FQDNNamespacePrefix\FQDN;
+
+class Foo
+{
+    public function bar()
+    {
+        $fqdn = new FQDN();
+        $someClass = new FQDN\SomeClass();
+    }
+}