--- a/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/Functions.php
+++ b/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/Functions.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace IgnoreNamespace\Deps;
+
+function functionUsedWithFQDN() {}
+
+function functionUsedWithNamespace() {}