--- a/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/UsedWithFQDN.php
+++ b/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/UsedWithFQDN.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace IgnoreNamespace\Deps;
+
+class UsedWithFQDN {}