--- a/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/UsedWithNamespace.php
+++ b/tests/DependencyEmitter/Fixtures/IgnoreNamespace/Deps/UsedWithNamespace.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace IgnoreNamespace\Deps;
+
+class UsedWithNamespace {}