--- a/tests/DependencyEmitter/UsesDependencyEmitterTest.php
+++ b/tests/DependencyEmitter/UsesDependencyEmitterTest.php
@@ -26,4 +26,37 @@ final class UsesDependencyEmitterTest extends TestCase
         self::assertCount(1, $deps);
         self::assertContains('Foo\Bar:4 on SomeUse', $deps);
     }
+
+    public function testIgnoresNamespaces(): void
+    {
+        $deps = $this->getDeps(
+            new UsesDependencyEmitter(),
+            [
+                __DIR__.'/Fixtures/IgnoreNamespace/Deps/UsedWithFQDN.php',
+                __DIR__.'/Fixtures/IgnoreNamespace/Deps/UsedWithNamespace.php',
+                __DIR__.'/Fixtures/IgnoreNamespace/Deps/Functions.php',
+                __DIR__.'/Fixtures/IgnoreNamespace/Uses/Foo.php',
+            ]
+        );
+
+        self::assertCount(2, $deps);
+        self::assertNotContains('IgnoreNamespace\Uses\Foo:5 on IgnoreNamespace\Deps', $deps);
+        self::assertContains('IgnoreNamespace\Uses\Foo:6 on IgnoreNamespace\Deps\UsedWithFQDN', $deps);
+        self::assertContains('IgnoreNamespace\Uses\Foo:7 on IgnoreNamespace\Deps\Functions\functionUsedWithFQDN', $deps);
+    }
+
+    public function testIncludesFQDNWhichIsAlsoANamespacePrefix(): void
+    {
+        $deps = $this->getDeps(
+            new UsesDependencyEmitter(),
+            [
+                __DIR__.'/Fixtures/FQDNNamespacePrefix/FQDN.php',
+                __DIR__.'/Fixtures/FQDNNamespacePrefix/FQDN/SomeClass.php',
+                __DIR__.'/Fixtures/FQDNNamespacePrefix/Uses/Foo.php',
+            ]
+        );
+
+        self::assertCount(1, $deps);
+        self::assertContains('FQDNNamespacePrefix\Uses\Foo:5 on FQDNNamespacePrefix\FQDN', $deps);
+    }
 }