Fixes #758.

Prevent namespaces from being added to use dependencies. Namespaces are
detected by looking if the used FQDN is a prefix of some class FQDN.

This strategy may be a bit brittle, I'll let you judge if this good enough or if the current behavior is better.