@mdevlamynck 2022-01-06
Yes I'll add tests :) Specifically the only issue I can foresee with this approach is considering that `Qossmic\Deptrac\AstRunner\AstMap` is a namespace (because it's a prefix of `Qossmic\Deptrac\AstRunner\AstMap\AstDependency` for instance) while it's also a FQDN. I'll definitely try to test that.