@pincher2012 2022-01-20
Hello, thank You for this tool and your contribution!

I just want to share performance results on my project before and after this PR.

Before:
```
 -------------------- -------- 
  Report                       
 -------------------- -------- 
  Violations           0       
  Skipped violations   901     
  Uncovered            11857   
  Allowed              106775  
  Warnings             0       
  Errors               0       
 -------------------- -------- 

real    0m8,483s
user    0m7,044s
sys     0m1,216s
```

After:
```
 -------------------- -------- 
  Report                       
 -------------------- -------- 
  Violations           0       
  Skipped violations   901     
  Uncovered            11464   
  Allowed              106775  
  Warnings             0       
  Errors               0       
 -------------------- -------- 


real    3m51,235s
user    3m49,009s
sys     0m1,367s
```

How do you think, is it possible to optimize algorythm?